Abstract relative date function into separate file to enable sharing
